feat: connect expense
