+++ Frontend/lib/api.ts
@@ -13,10 +13,21 @@ import {
   CreateGroupResponse,
   GetGroupsResponse,
   GetGroupByIdResponse,
+  GetGroupExpensesResponse,
+  SettlementResponse,
+  AddMemberResponse,
+  CreateExpenseRequest,
+  CreateExpenseResponse,
 } from "./types";
 
 // API Configuration
 const API_BASE_URL = "http://localhost:3001";
+
+// Extended Error type for API errors
+interface ApiErrorWithStatus extends Error {
+  statusCode?: number;
+  error?: string;
+}
 
 // Generic API request handler with JWT support
 async function apiRequest<T>(
@@ -63,9 +74,9 @@ async function apiRequest<T>(
       // Create a more descriptive error message
       const errorMessage =
         errorData.message || getErrorMessage(response.status);
-      const error = new Error(errorMessage);
-      (error as any).statusCode = errorData.statusCode || response.status;
-      (error as any).error = errorData.error;
+      const error = new Error(errorMessage) as ApiErrorWithStatus;
+      error.statusCode = errorData.statusCode || response.status;
+      error.error = errorData.error;
 
       throw error;
     }
@@ -138,12 +149,7 @@ export const authApi = {
     return response.json();
   },
 
-  async removeAvatar(): Promise<{ message: string }> {
-    return apiRequest<{ message: string }>("/user/remove-avatar", {
-      method: "PATCH",
-    });
-  },
-
+  // Upload payment QR code
   async uploadPaymentQr(
     file: File
   ): Promise<{ message: string; path: string }> {
@@ -205,22 +211,32 @@ export const groupApi = {
     });
   },
 
-  async getGroupExpenses(groupId: string): Promise<any> {
-    return apiRequest(`/groups/${groupId}/expenses`, {
+  async getGroupExpenses(groupId: string): Promise<GetGroupExpensesResponse> {
+    return apiRequest<GetGroupExpensesResponse>(`/groups/${groupId}/expenses`, {
       method: "GET",
     });
   },
 
-  async getSettlement(groupId: string): Promise<any> {
-    return apiRequest(`/groups/${groupId}/settlement`, {
+  async getSettlement(groupId: string): Promise<SettlementResponse> {
+    return apiRequest<SettlementResponse>(`/groups/${groupId}/settlement`, {
       method: "GET",
     });
   },
 
-  async addMember(groupId: string, userId: string): Promise<any> {
-    return apiRequest(`/groups/${groupId}/members`, {
+  async addMember(groupId: string, userId: string): Promise<AddMemberResponse> {
+    return apiRequest<AddMemberResponse>(`/groups/${groupId}/members`, {
       method: "POST",
       body: JSON.stringify({ userId }),
+    });
+  },
+};
+
+// Expense API functions
+export const expenseApi = {
+  async createExpense(data: CreateExpenseRequest): Promise<CreateExpenseResponse> {
+    return apiRequest<CreateExpenseResponse>("/expenses", {
+      method: "POST",
+      body: JSON.stringify(data),
     });
   },
 };
@@ -303,10 +319,11 @@ export const retryRequest = async <T>(
       lastError = error as Error;
 
       // Don't retry on client errors (4xx)
+      const errorWithStatus = lastError as ApiErrorWithStatus;
       if (
         lastError instanceof Error &&
-        (lastError as any).statusCode &&
-        (lastError as any).statusCode < 500
+        errorWithStatus.statusCode &&
+        errorWithStatus.statusCode < 500
       ) {
         throw lastError;
       }
@@ -326,10 +343,12 @@ export const retryRequest = async <T>(
   throw lastError!;
 };
 
+// Utility to get avatar URL
 export const getAvatarUrl = (userId: string): string => {
   return `${API_BASE_URL}/user/${userId}/avatar`;
 };
 
+// Utility to get payment QR URL
 export const getPaymentQrUrl = (userId: string): string => {
   return `${API_BASE_URL}/user/${userId}/payment-qr`;
 };
